+++ packages/eslint-plugin-stories/CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## Unreleased
 
+## 3.0.1 (2021-12-03)
+
+- [fix] Wrong buld output was published in v3.0.0
+
+## 3.0.0 (2021-12-03)
+
 - [breaking] Remove `no-jest-in-stories` [#22](https://github.com/chanzuckerberg/frontend-libs/pull/22)
 - [breaking] Remove `stories-default-export` [#22](https://github.com/chanzuckerberg/frontend-libs/pull/22)
 - [fix] Tweak failure message for `no-ext-resources-in-stories` [#22](https://github.com/chanzuckerberg/frontend-libs/pull/22)
 - [new] Add `no-csf-v2` rule [#22](https://github.com/chanzuckerberg/frontend-libs/pull/22)
-- [fix] Wrong buld output was published in v3.0.0
 
 ## 2.0.1 (2021-11-29)